deprecate helpers module in favor of vapo_lib
